Overhaul UI
Cleaned up the CSS to make the page more professional and readable.

Updated the tables on gameboard.html to contain a game state for the battleship game.

renamed contact.html to about.html (website does not require a contact page but does eventually require a page containing important details about the source code)